requirements.txt
asgiref
certifi
charset-normalizer
dj-database-url
Django
django-background-tasks
django-cors-headers
django-environ
django-filter
djangorestframework
idna
packaging
pillow
pip
pipenv
psycopg2-binary
requests
setuptools
six
sqlparse
sslcommerz-lib
tzdata
urllib3
virtualenv
whitenoise
